Review

+++ src/main/java/io/library/dao/BorrowBookDao.java
@@ -13,6 +13,7 @@ import java.sql.SQLException;
 import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class BorrowBookDao implements IBorrowBookDao {
 
@@ -25,19 +26,20 @@ public class BorrowBookDao implements IBorrowBookDao {
                 "       borrowed_books.return_date\n" +
                 "       FROM borrowed_books INNER JOIN books on borrowed_books.book_id = books.id INNER JOIN users on borrowed_books.user_id = users.id\n" +
                 "       WHERE user_id=%d", userId);
-        ResultSet resultSet = DataSourceDatabase.sqlExecutionerForSelect(sql);
-        while (resultSet.next()) {
-            String bookId = resultSet.getString("books.id");
-            String bookName = resultSet.getString("books.book_name");
-            String bookAuthor = resultSet.getString("books.book_author");
-            String borrowedDate = resultSet.getString("borrowed_books.borrowed_date");
-            String returnDate = resultSet.getString("borrowed_books.return_date");
-            String name = resultSet.getString("users.name");
-            String genre = resultSet.getString("books.genre");
-            int quantity = resultSet.getInt("books.quantity");
-            Book book = new Book(bookId, bookName, bookAuthor, quantity, genre);
-            BorrowedBook bookDao = new BorrowedBook(book, borrowedDate, returnDate, name);
-            borrowedBooks.add(bookDao);
+        try(ResultSet resultSet = DataSourceDatabase.sqlExecutionerForSelect(sql)) {
+            while (resultSet.next()) {
+                String bookId = resultSet.getString("books.id");
+                String bookName = resultSet.getString("books.book_name");
+                String bookAuthor = resultSet.getString("books.book_author");
+                String borrowedDate = resultSet.getString("borrowed_books.borrowed_date");
+                String returnDate = resultSet.getString("borrowed_books.return_date");
+                String name = resultSet.getString("users.name");
+                String genre = resultSet.getString("books.genre");
+                int quantity = resultSet.getInt("books.quantity");
+                Book book = new Book(bookId, bookName, bookAuthor, quantity, genre);
+                BorrowedBook bookDao = new BorrowedBook(book, borrowedDate, returnDate, name);
+                borrowedBooks.add(bookDao);
+            }
         }
         return borrowedBooks;
     }
@@ -46,7 +48,7 @@ public class BorrowBookDao implements IBorrowBookDao {
         return getAllBorrowedBook(userName).size();
     }
 
-    public void borrowABook(String bookId, String userName) {
+    public void borrowABook(String bookId, String userName) throws SQLException {
         try {
             DateTime today = new DateTime();
             DateTime returnDate = new DateTime().plusDays(7);
@@ -61,31 +63,37 @@ public class BorrowBookDao implements IBorrowBookDao {
             System.out.println(sql);
             DataSourceDatabase.sqlExecutionerForDML(sql);
             GlobalDataSource.getDataSource().getBookDao().decreaseQuantityOfBook(bookId, 1);
+            DataSourceDatabase.commitToDatabase();
         }
         catch (SQLIntegrityConstraintViolationException exception) {
-            System.out.println("Book doesn't exist");
+            DataSourceDatabase.rollbackUncommittedStatement();
+            throw new NoSuchElementException("There is no such book");
         }
         catch (SQLException exception) {
-            System.out.println("SQL Error");
-            exception.printStackTrace();
+            DataSourceDatabase.rollbackUncommittedStatement();
+            throw new SQLException("SQL can't be executed");
         }
     }
 
-    public void returnABook(String bookId, String userName) throws SQLException {
+    public void returnABook(String bookId, String userName) throws SQLException, IllegalStateException {
             int userId = GlobalDataSource.getDataSource().getUserDao().getUserId(userName);
             var borrowedBooks = getAllBorrowedBook(userName);
-            boolean returned = false;
+            if(borrowedBooks == null || borrowedBooks.isEmpty()) {
+                throw new IllegalStateException("No such borrowed book");
+            }
             for(var book : borrowedBooks) {
                 if(book.getBook().getId().equals(bookId)) {
-                    String sql = String.format("DELETE FROM borrowed_books WHERE user_id=%d AND book_id=%s",userId, Utility.getFormattedString(bookId));
-                    System.out.println(sql);
-                    DataSourceDatabase.sqlExecutionerForDML(sql);
-                    GlobalDataSource.getDataSource().getBookDao().increaseQuantityOfBook(bookId, 1);
-                    returned = true;
+                    try {
+                        String sql = String.format("DELETE FROM borrowed_books WHERE user_id=%d AND book_id=%s",userId, Utility.getFormattedString(bookId));
+                        DataSourceDatabase.sqlExecutionerForDML(sql);
+                        GlobalDataSource.getDataSource().getBookDao().increaseQuantityOfBook(bookId, 1);
+                        DataSourceDatabase.commitToDatabase();
+                    }
+                    catch (SQLException exception) {
+                        DataSourceDatabase.rollbackUncommittedStatement();
+                        throw new SQLException(exception.getMessage());
+                    }
                 }
-            }
-            if(!returned) {
-                throw new IllegalStateException("No such borrowed book");
             }
     }
 }
